requirements.txt
absl-py==1.0.0
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
convertdate==2.3.2
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
entrypoints==0.3
ephem==4.1.3
filelock==3.4.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
hijri-converter==2.2.2
idna==3.3
importlib-metadata==4.10.0
ipykernel==6.6.1
ipython==7.31.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.1
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyMeeus==0.5.11
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.6
pyspark==3.2.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools-git==1.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.29
sshtunnel==0.4.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
